refactor: MusicBar 로직 정리
